move predict to kaggle
